RT Rampage vdev ChainWalker

Documentation for ChainWalker.

Functions

set_charset(chain_walker_context, s_charset)

set_charset_opts(chain_walker_context, s_charset, n_min_length, n_max_length)

set_hash_routine(s_hash_routine)

Set the hash routine to use for hashing during operation.

Parameters

s_hash_routine: String to select the hashing algorithm to use.

Examples

iex> ChainWalker.set_hash_routine("md5")
%ChainWalkerContext{charset: nil, hash_routine: :md5, max_length: nil, min_length: nil, table_index: nil}

set_max_password_length(chain_walker_context, n_max_length)

set_min_password_length(chain_walker_context, n_min_length, n_max_length)

Set minimum plain text password length to be included in the rainbow table.

Parameters

chain_walker_context: %ChainWalkerContext{}

n_min_length: Minimum length for plain text passwords.

Returns

%ChainWalkerContext{}

Examples

iex> ChainWalker.set_min_password_length(%ChainWalkerContext{}, 1, 2)
%ChainWalkerContext{charset: nil, hash_routine: nil, max_length: nil, min_length: 1, table_index: nil}

set_table_index(chain_walker_context, n_table_index)

validate_arg(struct, argument)

Validates a given argument, by checking if it exists as a key in the struct provided.

Parameters

struct: Struct containing valid options and their values.

argument: Argument entered to be validated.

Examples

iex> ChainWalker.validate_arg(%HashRoutines{}, "md5")
{:ok, :md5}
iex> ChainWalker.validate_arg(%HashRoutines{}, "unknown_routine")
{:error, "Incorrect argument 'unknown_routine', is not supported!"}

validate_min_password_length(n_min_length, n_max_length)

Validates correctness of minimum password length argument.

Parameters

n_min_length: Minimum plain text password length.

n_max_length: Maximum plain text password length.

Returns

{ :ok, n_min_length }

{ :error, reason }

Examples

iex> ChainWalker.validate_min_password_length(1, 2)
{:ok, 1}
iex> ChainWalker.validate_min_password_length(2, 1)
{:error, "Bad argument: Minimum password length (2) cannot be greater than maximum length (1)."}
iex> ChainWalker.validate_min_password_length(0, 2)
{:error, "Bad argument: Mininum password length (0) cannot be zero or less."}
iex> ChainWalker.validate_min_password_length(10, 10)
{:error, "Bad argument: Minimum password length (10) cannot be greater than maximum supported password length (9)." }

validate_routine(s_hash_routine)

Validates the given hash_algorithm argument, s_hash_routine by checking if the hash algorithm is supported.

Parameters

s_hash_routine: String to select the hashing algorithm to use.

Examples

iex> ChainWalker.validate_routine("md5")
{:ok, :md5}